% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY.R
\name{calc_omega}
\alias{calc_omega}
\title{Calculate Omega matrices}
\usage{
calc_omega(eval, D_l)
}
\arguments{
\item{eval}{vector of eigenvalues from decomposition of relatedness matrix}

\item{D_l}{vector of length d_size}
}
\description{
Calculate Omega matrices
}
