% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY.R
\name{calc_qi}
\alias{calc_qi}
\title{Calculate Qi (inverse of Q) and log determinant of Q}
\usage{
calc_qi(eval, D_l, X)
}
\arguments{
\item{eval}{vector of eigenvalues from decomposition of relatedness matrix}

\item{D_l}{vector of length d_size}

\item{X}{a design matrix}
}
\description{
Calculate Qi (inverse of Q) and log determinant of Q
}
