% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_qi2.R
\name{calc_qi2}
\alias{calc_qi2}
\title{Calculate Qi (inverse of Q) and log determinant of Q for 2 loci (which might be the same locus)}
\usage{
calc_qi2(eval, D_l, X)
}
\arguments{
\item{eval}{vector of eigenvalues from decomposition of relatedness matrix}

\item{D_l}{eigenvalues vector of length d_size}

\item{X}{a 16 by n design matrix with genotype probabilities for two loci}
}
\description{
Calculate Qi (inverse of Q) and log determinant of Q for 2 loci (which might be the same locus)
}
