% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_qi_2loci.R
\name{calc_qi_2loci}
\alias{calc_qi_2loci}
\title{Calculate Qi (inverse of Q) and log determinant of Q for two loci}
\usage{
calc_qi_2loci(eval, D_l, X1, X2)
}
\arguments{
\item{eval}{vector of eigenvalues from decomposition of relatedness matrix}

\item{D_l}{vector of length d_size}

\item{X1}{a design matrix}

\item{X2}{a design matrix}
}
\description{
Calculate Qi (inverse of Q) and log determinant of Q for two loci
}
