% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{calc_sigma}
\alias{calc_sigma}
\title{Calculate Sigma_ee and Sigma_uu matrices}
\usage{
calc_sigma(eval, D_l, X, OmegaU, OmegaE, UltVeh, Qi)
}
\arguments{
\item{eval}{eigenvalues vector from decomposition of relatedness matrix}

\item{D_l}{vector}

\item{X}{design matrix}

\item{OmegaU}{matrix}

\item{OmegaE}{matrix}

\item{UltVeh}{matrix}

\item{Qi}{inverse of Q matrix}
}
\description{
Calculate Sigma_ee and Sigma_uu matrices
}
