% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sigma_2loci.R
\name{calc_sigma_2loci}
\alias{calc_sigma_2loci}
\title{Calculate Sigma_ee and Sigma_uu matrices for 2 loci}
\usage{
calc_sigma_2loci(eval, D_l, X1, X2, OmegaU, OmegaE, UltVeh, Qi)
}
\arguments{
\item{eval}{eigenvalues vector from decomposition of relatedness matrix}

\item{D_l}{vector}

\item{X1}{design matrix}

\item{X2}{design matrix}

\item{OmegaU}{matrix}

\item{OmegaE}{matrix}

\item{UltVeh}{matrix}

\item{Qi}{inverse of Q matrix}
}
\description{
Calculate Sigma_ee and Sigma_uu matrices for 2 loci
}
