% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY.R
\name{calc_XHiY}
\alias{calc_XHiY}
\title{Calculate XHiY}
\usage{
calc_XHiY(eval, D_l, X, UltVehiY)
}
\arguments{
\item{eval}{vector of eigenvalues from the decomposition of the relatedness matrix}

\item{D_l}{vector of length d_size}

\item{X}{design matrix}

\item{UltVehiY}{a matrix}
}
\description{
Calculate XHiY
}
