% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY_2loci.R
\name{calc_XHiY_2loci}
\alias{calc_XHiY_2loci}
\title{Calculate XHiY for two loci}
\usage{
calc_XHiY_2loci(eval, D_l, X1, X2, UltVehiY)
}
\arguments{
\item{eval}{vector of eigenvalues from the decomposition of the relatedness matrix}

\item{D_l}{vector of length d_size}

\item{X1}{design matrix for locus 1}

\item{X2}{design matrix for locus 2}

\item{UltVehiY}{a matrix}
}
\description{
Calculate XHiY for two loci
}
