% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_kinship.R
\name{center_kinship}
\alias{center_kinship}
\title{Center a relatedness matrix, after Zhou's GEMMA function CenterMatrix}
\usage{
center_kinship(mat)
}
\arguments{
\item{mat}{a relatedness matrix}
}
\description{
Center a relatedness matrix, after Zhou's GEMMA function CenterMatrix
}
\examples{
readr::read_tsv(system.file("extdata", "mouse100.cXX.txt", package = "gemma2"), col_names = FALSE)[, 1:100] -> kinship
e_out <- eigen2(as.matrix(kinship))
center_kinship(as.matrix(kinship)) -> kinship_centered
}
