% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_matrix.R
\name{center_matrix}
\alias{center_matrix}
\title{Center a kinship matrix before eigendecomposition}
\usage{
center_matrix(mat)
}
\arguments{
\item{mat}{a kinship matrix}
}
\description{
Center a kinship matrix before eigendecomposition
}
