% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen2.R
\name{eigen2}
\alias{eigen2}
\title{Calculate eigendecomposition and return ordered eigenvalues and eigenvectors}
\usage{
eigen2(spd, decreasing = FALSE)
}
\arguments{
\item{spd}{a semi-positive definite matrix}

\item{decreasing}{argument passed to order()}
}
\description{
Calculate eigendecomposition and return ordered eigenvalues and eigenvectors
}
