% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY.R
\name{eigen_proc}
\alias{eigen_proc}
\title{Eigendecomposition procedure for Vg and Ve}
\usage{
eigen_proc(V_g, V_e)
}
\arguments{
\item{V_g}{a d_size by d_size covariance matrix}

\item{V_e}{a d_size by d_size covariance matrix}
}
\description{
Eigendecomposition procedure for Vg and Ve
}
