% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MphCalcLogL.R
\name{MphCalcLogL}
\alias{MphCalcLogL}
\title{Calculate log likelihood}
\usage{
MphCalcLogL(eval, D_l, Qi, UltVehiY, xHiy)
}
\arguments{
\item{eval}{eigenvalues vector from decomposition of relatedness matrix}

\item{D_l}{vector}

\item{Qi}{inverse of Q matrix}

\item{UltVehiY}{matrix of (transformed) Y values}

\item{xHiy}{vector}
}
\description{
Calculate log likelihood
}
