% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MphCalcLogL.R
\name{MphEM}
\alias{MphEM}
\title{Perform EM algorithm}
\usage{
MphEM(max_iter = 10000, max_prec = 1/1000, eval, X, Y, V_g, V_e)
}
\arguments{
\item{max_iter}{maximum number of iterations for EM algorithm}

\item{max_prec}{maximum precision for EM algorithm}

\item{eval}{vector of eigenvalues from relatedness matrix decomposition}

\item{X}{design matrix}

\item{Y}{matrix of phenotype values}

\item{V_g}{covariance matrix}

\item{V_e}{covariance matrix}
}
\description{
Perform EM algorithm
}
