% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MphEM_2loci.R
\name{MphEM_2loci}
\alias{MphEM_2loci}
\title{Perform EM algorithm for pleiotropy or close linkage}
\usage{
MphEM_2loci(max_iter = 10000, max_prec = 1/1000, eval, X1, X2, Y,
  V_g = diag(1, 2), V_e = diag(1, 2))
}
\arguments{
\item{max_iter}{maximum number of iterations for EM algorithm}

\item{max_prec}{maximum precision for EM algorithm}

\item{eval}{vector of eigenvalues from relatedness matrix decomposition}

\item{X1}{transformed genotype probabilities matrix 1}

\item{X2}{transformed genotype probabilities matrix 2}

\item{Y}{transformed matrix of phenotype values}

\item{V_g}{genetic covariance matrix}

\item{V_e}{error covariance matrix}
}
\description{
Perform EM algorithm for pleiotropy or close linkage
}
