% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagger_mats.R
\name{stagger_mats}
\alias{stagger_mats}
\title{Stagger matrices of same size within a larger, block-diagonal matrix}
\usage{
stagger_mats(...)
}
\arguments{
\item{X1}{first matrix}

\item{X2}{second matrix, of same size as first}
}
\description{
Stagger matrices of same size within a larger, block-diagonal matrix
}
