% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{update_e}
\alias{update_e}
\title{Update E}
\usage{
update_e(UltVehiY, UltVehiBX, UltVehiU)
}
\arguments{
\item{UltVehiY}{matrix of transformed Y values}

\item{UltVehiBX}{matrix of transformed BX values}

\item{UltVehiU}{matrix of transformed U values}
}
\description{
Update E
}
