% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{update_v}
\alias{update_v}
\title{Update V_e and V_g}
\usage{
update_v(eval, U, E, Sigma_uu, Sigma_ee)
}
\arguments{
\item{eval}{vector of eigenvalues from eigendecomposition of relatedness matrix}

\item{U}{matrix}

\item{E}{matrix}

\item{Sigma_uu}{matrix}

\item{Sigma_ee}{matrix}

\item{tol}{a positive number indicating tolerance to be passed to isSymmetric()}
}
\description{
Update V_e and V_g
}
