% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{UpdateRL_B}
\alias{UpdateRL_B}
\title{Update B for restricted log likelihood}
\usage{
UpdateRL_B(xHiy, Qi, d_size)
}
\arguments{
\item{xHiy}{vector}

\item{Qi}{Q inverse matrix}

\item{d_size}{number of traits}
}
\description{
Update B for restricted log likelihood
}
\seealso{
Other expectation-maximization functions: 
\code{\link{update_e}()},
\code{\link{update_u}()},
\code{\link{update_v}()}
}
\concept{expectation-maximization functions}
