% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen2.R
\name{eigen2}
\alias{eigen2}
\title{Calculate eigendecomposition and return ordered eigenvalues and eigenvectors}
\usage{
eigen2(spd, decreasing = FALSE)
}
\arguments{
\item{spd}{a semi-positive definite matrix}

\item{decreasing}{argument passed to order()}
}
\value{
a list with 2 components, the eigenvalues and the eigenvectors
}
\description{
Calculate eigendecomposition and return ordered eigenvalues and eigenvectors
}
\examples{
readr::read_tsv(system.file("extdata", "mouse100.cXX.txt", package = "gemma2"), col_names = FALSE)[, 1:100] -> kinship
e_out <- eigen2(as.matrix(kinship))
}
