% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY.R
\name{eigen_proc}
\alias{eigen_proc}
\title{Eigendecomposition procedure for Vg and Ve}
\usage{
eigen_proc(V_g, V_e, tol = 1/10000)
}
\arguments{
\item{V_g}{a d_size by d_size covariance matrix}

\item{V_e}{a d_size by d_size covariance matrix}

\item{tol}{a positive number indicating the tolerance for isSymmetric}
}
\value{
a named list of length 4 containing the outputs of eigendecomposition procedure
}
\description{
Eigendecomposition procedure for Vg and Ve
}
\examples{
eigen_proc(diag(2), diag(2))

}
