% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_XHiY.R
\name{eigen_proc}
\alias{eigen_proc}
\title{Eigendecomposition procedure for Vg and Ve}
\usage{
eigen_proc(Vg, Ve)
}
\arguments{
\item{Vg}{a d_size by d_size covariance matrix}

\item{Ve}{a d_size by d_size covariance matrix}
}
\description{
Eigendecomposition procedure for Vg and Ve
}
