% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MphEM.R
\name{MphEM}
\alias{MphEM}
\title{Perform EM algorithm}
\usage{
MphEM(max_iter = 10000, max_prec = 1/1e+06, eval, X, Y, V_g, V_e)
}
\arguments{
\item{max_iter}{maximum number of iterations for EM algorithm}

\item{max_prec}{maximum precision for EM algorithm}

\item{eval}{vector of eigenvalues from relatedness matrix decomposition}

\item{X}{design matrix}

\item{Y}{matrix of phenotype values}

\item{V_g}{genetic covariance matrix}

\item{V_e}{error covariance matrix}
}
\value{
a list of lists. Length of list corresponds to number of EM iterations
}
\description{
Perform EM algorithm
}
