% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagger_mats.R
\name{stagger_mats}
\alias{stagger_mats}
\title{Stagger matrices within a larger, block-diagonal matrix}
\usage{
stagger_mats(...)
}
\arguments{
\item{...}{one or more matrices, separated by commas}
}
\description{
Stagger matrices within a larger, block-diagonal matrix
}
