% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MphCalcLogL.R
\name{stagger_mats}
\alias{stagger_mats}
\title{Stagger two matrices of same size within a larger, block-diagonal matrix}
\usage{
stagger_mats(X1, X2)
}
\arguments{
\item{X1}{first matrix}

\item{X2}{second matrix, of same size as first}
}
\description{
Stagger two matrices of same size within a larger, block-diagonal matrix
}
