% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagger_mats.R
\name{stagger_mats}
\alias{stagger_mats}
\title{Stagger matrices within a larger, block-diagonal matrix}
\usage{
stagger_mats(...)
}
\arguments{
\item{...}{one or more matrices, separated by commas}
}
\value{
a block-diagonal matrix, with the inputted matrices as blocks on the diagonal.
}
\description{
Stagger matrices within a larger, block-diagonal matrix
}
\examples{
foo <- matrix(rnorm(40000), ncol = 8)
block_diag <- stagger_mats(foo, foo)
dim(foo)
dim(block_diag)
}
