% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{update_e}
\alias{update_e}
\title{Update E}
\usage{
update_e(UltVehiY, UltVehiBX, UltVehiU)
}
\arguments{
\item{UltVehiY}{matrix of transformed Y values}

\item{UltVehiBX}{matrix of transformed BX values}

\item{UltVehiU}{matrix of transformed U values}
}
\description{
Update E
}
\seealso{
Other expectation-maximization functions: 
\code{\link{UpdateRL_B}()},
\code{\link{update_u}()},
\code{\link{update_v}()}
}
\concept{expectation-maximization functions}
