% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{update_u}
\alias{update_u}
\title{Update U matrix}
\usage{
update_u(OmegaE, UltVehiY, UltVehiBX)
}
\arguments{
\item{OmegaE}{the OmegaE matrix, calculated in calc_omega}

\item{UltVehiY}{matrix}

\item{UltVehiBX}{matrix}
}
\description{
Update U matrix
}
