% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_u.R
\name{UpdateRL_B}
\alias{UpdateRL_B}
\title{Update restricted log likelihood}
\usage{
UpdateRL_B(xHiy, Qi, d_size = 2)
}
\arguments{
\item{xHiy}{vector}

\item{Qi}{Q inverse matrix}

\item{d_size}{number of traits}
}
\description{
Update restricted log likelihood
}
